+++ src/import/index.ts
@@ -1,3 +1,4 @@
 /*! Copyright (c) 2022, XAPP AI*/
 export * from "./importApp";
 export * from "./importFromDialogflow";
+export * from "./importAppFromFile";
